refactor: refactor getUserOrGroupWithRoleInfo
